backend/requirements.txt
fastapi[all]
SQLAlchemy
uvicorn[standard]
pydantic-settings
PyMySQL       # ★この行を追加★
cryptography